+++ apps/backend/fixtures/lambda/daily-filecheck.json
@@ -1,0 +1,1 @@
+{}
