+++ apps/backend/fixtures/lambda/parser.json
@@ -1,1 +1,1 @@
-{ "eventType": "all" }
+{ "Records": [] }
