+++ apps/backend/package.json
@@ -47,9 +47,9 @@
     "decimal.js": "10.4.3",
     "exceljs": "4.3.0",
     "express": "4.18.2",
+    "nats": "2.15.1",
     "nest-aws-sdk": "3.0.0",
     "nest-winston": "1.8.0",
-    "node-nats-streaming": "0.3.2",
     "pg": "8.8.0",
     "reflect-metadata": "0.1.13",
     "rimraf": "4.1.1",
@@ -57,8 +57,7 @@
     "typeorm": "0.3.11",
     "underscore": "1.13.6",
     "webpack": "5.75.0",
-    "winston": "3.8.2",
-    "winston-daily-rotate-file": "4.7.1"
+    "winston": "3.8.2"
   },
   "devDependencies": {
     "@compodoc/compodoc": "1.1.19",
@@ -77,21 +76,17 @@
     "@types/underscore": "1.11.4",
     "@typescript-eslint/eslint-plugin": "5.48.2",
     "@typescript-eslint/parser": "5.48.2",
-    "aws-sdk": "2.1299.0",
     "eslint": "8.32.0",
     "eslint-config-prettier": "8.6.0",
     "eslint-nibble": "8.1.0",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-unused-imports": "2.0.0",
-    "import-local": "3.1.0",
     "jest": "29.3.1",
     "lint-staged": "13.1.0",
     "prettier": "2.8.3",
-    "source-map-support": "0.5.21",
     "supertest": "6.3.3",
     "ts-jest": "29.0.5",
-    "ts-loader": "9.4.2",
     "ts-node": "10.9.1",
     "tsconfig-paths": "4.1.2",
     "typescript": "4.9.4"
