+++ apps/backend/src/lambdas/reconcile.ts
@@ -9,7 +9,7 @@ import { NormalizedLocation } from '../constants';
 import { CashDepositService } from '../deposits/cash-deposit.service';
 import { PosDepositService } from '../deposits/pos-deposit.service';
 import { LocationService } from '../location/location.service';
-import { ParseService } from '../parse/parse.service';
+import { NotificationService } from '../notification/notification.service';
 import { CashExceptionsService } from '../reconciliation/cash-exceptions.service';
 import { CashReconciliationService } from '../reconciliation/cash-reconciliation.service';
 import { PosReconciliationService } from '../reconciliation/pos-reconciliation.service';
@@ -32,7 +32,7 @@ export const handler = async (event: ReconciliationConfigInput) => {
   const posDepositService = app.get(PosDepositService);
   const locationService = app.get(LocationService);
   const reportingService = app.get(ReportingService);
-  const parseService = app.get(ParseService);
+  const notificationService = app.get(NotificationService);
   const appLogger = app.get(Logger);
   const currentDate = new Date();
   const numDaysToReconcile = differenceInDays(
@@ -48,12 +48,15 @@ export const handler = async (event: ReconciliationConfigInput) => {
 
   if (!event.bypass_parse_validity) {
     // Prevent reconciler from running for a program if no valid files today
-    const rule = await parseService.getRulesForProgram(event.program);
+    const rule = await notificationService.getRulesForProgram(event.program);
     if (!rule) {
       throw new Error('No rule for this program');
     }
     const reconciliationDate = new Date();
-    const daily = await parseService.getDailyForRule(rule, reconciliationDate);
+    const daily = await notificationService.getDailyForRule(
+      rule,
+      reconciliationDate
+    );
     if (!daily?.success) {
       throw new Error(
         `Incomplete dataset for this date ${reconciliationDate}. Please check the uploaded files.`
