+++ apps/backend/src/notification/entities/program-daily-upload.entity.ts
@@ -9,7 +9,7 @@ import {
   Relation,
 } from 'typeorm';
 import { FileIngestionRulesEntity } from './file-ingestion-rules.entity';
-import { FileUploadedEntity } from './file-uploaded.entity';
+import { FileUploadedEntity } from '../../parse/entities/file-uploaded.entity';
 
 @Entity('program_daily_upload')
 export class ProgramDailyUploadEntity {
