+++ apps/backend/src/parse/entities/file-uploaded.entity.ts
@@ -7,8 +7,9 @@ import {
   JoinColumn,
   Relation,
 } from 'typeorm';
-import { ProgramDailyUploadEntity } from './program-daily-upload.entity';
 import { FileTypes } from '../../constants';
+import { ProgramDailyUploadEntity } from '../../notification/entities/program-daily-upload.entity';
+import { ProgramDailyUploadEntity } from '../../notification/entities/program-daily-upload.entity';
 
 @Entity('file_uploaded')
 export class FileUploadedEntity {
@@ -32,7 +33,7 @@ export class FileUploadedEntity {
 
   @ManyToOne(
     () => ProgramDailyUploadEntity,
-    (programDailyUpload) => programDailyUpload.files
+    (programDailyUpload: ProgramDailyUploadEntity) => programDailyUpload.files
   )
   @JoinColumn({ name: 'daily_upload_id' })
   dailyUpload: Relation<ProgramDailyUploadEntity>;
