+++ apps/backend/src/parse/parse.service.ts
@@ -1,27 +1,41 @@
-import { Inject, Injectable, Logger } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpException,
+  HttpStatus,
+  Inject,
+  Injectable,
+  Logger,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { validateOrReject, ValidationError } from 'class-validator';
-import { format } from 'date-fns';
 import { Repository } from 'typeorm';
+import _ from 'underscore';
 import { CashDepositDTO, CashDepositsListDTO } from './dto/cash-deposit.dto';
 import {
   GarmsTransactionDTO,
   GarmsTransactionList,
 } from './dto/garms-transaction.dto';
 import { PosDepositDTO, PosDepositListDTO } from './dto/pos-deposit.dto';
-import { FileIngestionRulesEntity } from './entities/file-ingestion-rules.entity';
 import { FileUploadedEntity } from './entities/file-uploaded.entity';
-import { ProgramDailyUploadEntity } from './entities/program-daily-upload.entity';
 import { FileTypes, ParseArgsTDI } from '../constants';
+import { CashDepositService } from '../deposits/cash-deposit.service';
 import { CashDepositEntity } from '../deposits/entities/cash-deposit.entity';
 import { POSDepositEntity } from '../deposits/entities/pos-deposit.entity';
+import { PosDepositService } from '../deposits/pos-deposit.service';
 import { TDI17Details, TDI34Details } from '../flat-files';
+import { ParseEvent } from '../lambdas/interface';
 import { parseGarms } from '../lambdas/utils/parseGarms';
 import { parseTDI } from '../lambdas/utils/parseTDI';
 import { AppLogger } from '../logger/logger.service';
+import { FileIngestionRulesEntity } from '../notification/entities/file-ingestion-rules.entity';
+import { MAIL_TEMPLATE_ENUM } from '../notification/mail-templates';
+import { MailService } from '../notification/mail.service';
+import { NotificationService } from '../notification/notification.service';
+import { S3ManagerService } from '../s3-manager/s3-manager.service';
 import { TransactionEntity } from '../transaction/entities';
 import { SBCGarmsJson } from '../transaction/interface';
 import { PaymentMethodService } from '../transaction/payment-method.service';
+import { TransactionService } from '../transaction/transaction.service';
 
 @Injectable()
 export class ParseService {
@@ -29,50 +43,21 @@ export class ParseService {
     @Inject(Logger) private readonly appLogger: AppLogger,
     @Inject(PaymentMethodService)
     private readonly paymentMethodService: PaymentMethodService,
+    @Inject(S3ManagerService)
+    private readonly s3: S3ManagerService,
+    @Inject(PosDepositService)
+    private readonly posDepositService: PosDepositService,
+    @Inject(CashDepositService)
+    private readonly cashDepositService: CashDepositService,
+    @Inject(TransactionService)
+    private readonly transactionService: TransactionService,
+    @Inject(MailService)
+    private readonly mailService: MailService,
+    @Inject(NotificationService)
+    private readonly alertService: NotificationService,
     @InjectRepository(FileUploadedEntity)
-    private uploadedRepo: Repository<FileUploadedEntity>,
-    @InjectRepository(FileIngestionRulesEntity)
-    private ingestionRulesRepo: Repository<FileIngestionRulesEntity>,
-    @InjectRepository(ProgramDailyUploadEntity)
-    private programDailyRepo: Repository<ProgramDailyUploadEntity>
+    private uploadedRepo: Repository<FileUploadedEntity>
   ) {}
-
-  /**
-   * Determines if a daily status is successful or not based on the incoming rules
-   * @param rule
-   * @param files Files uploaded on the day
-   * @returns Success or no
-   */
-  determineDailySuccess(
-    rule: FileIngestionRulesEntity,
-    files: FileUploadedEntity[]
-  ): {
-    success: boolean;
-    missingFiles: { filename: string; fileType: FileTypes }[];
-  } {
-    const requiredFiles = rule.requiredFiles;
-    const missingFiles: { filename: string; fileType: FileTypes }[] = [];
-    requiredFiles.forEach((requiredFile) => {
-      if (
-        !files.some(
-          (file) =>
-            file.sourceFileType === requiredFile.fileType &&
-            file.sourceFileName.includes(requiredFile.filename)
-        )
-      ) {
-        missingFiles.push({
-          filename: requiredFile.filename,
-          fileType: requiredFile.fileType,
-        });
-      }
-    });
-    const success = missingFiles.length === 0;
-    return {
-      success,
-      missingFiles,
-    };
-  }
-
   /**
    * Handles validation errors from validateOrReject, identifying which property is failing validation
    * Returns the error message generated
@@ -121,7 +106,13 @@ export class ParseService {
     fileContents,
   }: ParseArgsTDI): Promise<unknown> {
     try {
-      return parseTDI({ type, fileName, program, fileContents });
+      //TODO upload? Or parse to the db?
+      return parseTDI({
+        type,
+        fileName,
+        program,
+        fileContents: Buffer.from(fileContents || '').toString(),
+      });
     } catch (err) {
       this.appLogger.error(err, 'Error parsing file');
       throw err;
@@ -140,7 +131,7 @@ export class ParseService {
   ): Promise<TransactionEntity[]> {
     const paymentMethods = await this.paymentMethodService.getPaymentMethods();
     // Creates an array of Transaction Entities
-    const garmsSales = await parseGarms(
+    const garmsSales = parseGarms(
       (await JSON.parse(contents || '{}')) as SBCGarmsJson[],
       fileName,
       paymentMethods
@@ -248,7 +239,7 @@ export class ParseService {
    * @returns List of uploaded files
    */
   async getAllFiles(): Promise<FileUploadedEntity[]> {
-    return this.uploadedRepo.find();
+    return await this.uploadedRepo.find();
   }
 
   /**
@@ -260,87 +251,257 @@ export class ParseService {
   async saveFileUploaded(
     fileUploaded: Partial<FileUploadedEntity>
   ): Promise<FileUploadedEntity> {
-    return this.uploadedRepo.save(fileUploaded);
+    return await this.uploadedRepo.save(fileUploaded);
   }
 
-  /**
-   * Gets all existing rules for each program
-   * @returns List of Rules
-   */
-  async getAllRules(): Promise<FileIngestionRulesEntity[]> {
-    return this.ingestionRulesRepo.find({
-      relations: {
-        requiredFiles: true,
-      },
-    });
-  }
+  async processAllFiles(event: ParseEvent) {
+    event.Records.length === 0
+      ? this.appLogger.log(
+          'No Records in Event...checking for files in bucket...'
+        )
+      : this.appLogger.log(`Processing ${event.Records.length} files...`);
 
-  /**
-   * Gets a single rule for a specific program
-   * @param program SBC or LABOUR for now
-   * @returns A set of rules for a program, or failure
-   */
-  async getRulesForProgram(program: string): Promise<FileIngestionRulesEntity> {
-    return this.ingestionRulesRepo.findOneOrFail({
-      where: {
-        program,
-      },
-    });
-  }
+    const eventFileList = event.Records.map((r) => r.s3.object.key);
 
-  /**
-   * Gets the daily status for a specified date for the specified rule
-   * @param rule
-   * @param date
-   * @returns A daily upload entity or nothing
-   */
-  async getDailyForRule(
-    rule: FileIngestionRulesEntity,
-    date: Date
-  ): Promise<ProgramDailyUploadEntity | null> {
-    return this.programDailyRepo.findOne({
-      relations: ['rule', 'files'],
-      where: {
-        dailyDate: format(date, 'yyyy-MM-dd'),
-        rule: {
-          id: rule.id,
-        },
-      },
-    });
-  }
+    const fileList =
+      (await this.s3.listBucketContents(
+        `pcc-integration-data-files-${process.env.RUNTIME_ENV}`
+      )) || [];
 
-  /**
-   * Creates a new daily entity for the specified rule for a specified date
-   * @param rule
-   * @param date
-   * @returns
-   */
-  async createNewDaily(
-    rule: FileIngestionRulesEntity,
-    date: Date
-  ): Promise<ProgramDailyUploadEntity> {
+    this.appLogger.log(
+      `Found ${fileList ? fileList.length : 0} files in bucket...`
+    );
+
     try {
-      const newDaily: Partial<ProgramDailyUploadEntity> = {
-        dailyDate: format(date, 'yyyy-MM-dd'),
-        success: false,
-        retries: 0,
-        rule,
-      };
-      const daily = this.programDailyRepo.create(newDaily);
-      return this.saveDaily(daily);
-    } catch (e) {
-      throw new Error('Error saving daily upload');
+      const allFiles = await this.getAllFiles();
+
+      this.appLogger.log(
+        `Found ${allFiles.length} files already uploaded to the  database...`
+      );
+
+      const allUploadedFiles: string[] = allFiles.map((f) => f.sourceFileName);
+
+      const parseList =
+        event.Records.length === 0
+          ? _.difference(fileList, allUploadedFiles)
+          : _.difference(eventFileList, allUploadedFiles);
+
+      this.appLogger.log(`Found ${parseList.length} files to parse...`);
+      // TODO [CCFPCM-318] Excluded LABOUR2 files that are DDF, needs implementation
+      const finalParseList = parseList.filter(
+        (filename) => !filename?.includes('LABOUR2')
+      );
+      this.appLogger.log('Creating daily upload for today if needed');
+
+      await this.commenceDailyUpload(new Date());
+      // Parse & Save only files that have not been parsed before
+      for (const filename of finalParseList) {
+        this.appLogger.log(`Parsing ${filename}..`);
+        if (filename) await this.processEvent(filename);
+      }
+    } catch (err) {
+      this.appLogger.error(err);
+    }
+  }
+
+  async processEvent(key: string) {
+    try {
+      const bucket = `pcc-integration-data-files-${process.env.RUNTIME_ENV}`;
+      const program = key.split('/')[0];
+      const filename = key.split('/')[1];
+      console.log(bucket, program, filename);
+
+      const file = await this.s3.getObject(bucket, `${program}/${filename}`);
+
+      let currentRule: FileIngestionRulesEntity | null = null;
+
+      const rules = await this.alertService.getAllRules();
+      const ministry = (() => {
+        for (const rule of rules) {
+          if (filename.includes(rule.program)) {
+            currentRule = rule;
+            return rule.program;
+          }
+        }
+        throw new Error(`File does not reference to any programs: ${filename}`);
+      })();
+
+      if (!currentRule) {
+        throw new Error('No rule associated');
+      }
+
+      const fileType = (() => {
+        const requiredFiles = currentRule?.requiredFiles;
+        const requiredFile = requiredFiles?.find((rf) =>
+          filename.includes(rf.filename)
+        );
+        if (!!requiredFile) {
+          return requiredFile.fileType;
+        }
+        throw new Error('Unknown file type: ' + filename);
+      })();
+
+      try {
+        this.appLogger.log('Call endpoint to upload file...', filename);
+        await this.uploadAndParseFile(
+          { fileName: `${program}/${filename}`, fileType, program: ministry },
+          Buffer.from(file.Body?.toString() || '')
+        );
+        // const formData = new FormData();
+        // formData.append('file', Readable.from(file), filename);
+        // formData.append('fileName', filename);
+        // formData.append('fileType', fileType);
+        // formData.append('program', ministry);
+        // await axiosInstance.post('/v1/parse/upload-file', formData, {
+        //   headers: {
+        //     ...formData.getHeaders(),
+        //   },
+        // });
+      } catch (err) {
+        this.appLogger.log('\n\n=========Errors with File Upload: =========\n');
+        this.appLogger.error(`Error with uploading file ${filename}`);
+        const errorMessage =
+          err instanceof BadRequestException
+            ? `Validation Errors in file ${filename}: ${err}`
+            : `Validation Errors present in the file ${filename}`;
+        this.appLogger.error(errorMessage);
+        const alertDestinations = await this.mailService.getAlertDestinations(
+          ministry,
+          [filename]
+        );
+        await this.mailService.sendEmailAlertBulk(
+          MAIL_TEMPLATE_ENUM.FILE_VALIDATION_ALERT,
+          alertDestinations.map((ad) => ({
+            toEmail: ad,
+            message: errorMessage,
+          }))
+        );
+      }
+    } catch (err) {
+      this.appLogger.error(err);
     }
   }
 
   /**
-   * Saves an existing daily with presumably new information
-   * @param daily Daily with updated information
-   * @returns ProgramDailyUploadEntity
+   * The below three functions are lifted straight from the parse controller
+   * This is to ensure its working within our lambda flows as the API Gateway
+   * is currently unable to take requests from the parsing lambda
    */
-  async saveDaily(
-    daily: ProgramDailyUploadEntity
-  ): Promise<ProgramDailyUploadEntity> {
-    return this.programDailyRepo.save(daily);
+  async commenceDailyUpload(date: Date) {
+    const rules = await this.alertService.getAllRules();
+    for (const rule of rules) {
+      const daily = await this.alertService.getDailyForRule(
+        rule,
+        new Date(date)
+      );
+      if (!daily) {
+        await this.alertService.createNewDaily(rule, new Date(date));
+      }
+    }
+  }
+
+  async uploadAndParseFile(
+    body: { fileName: string; fileType: FileTypes; program: string },
+    buffer: Buffer
+  ) {
+    const { fileName, fileType, program } = body;
+    this.appLogger.log(`Parsing ${fileName} - ${fileType}`);
+    const contents = buffer.toString();
+
+    const allFiles = await this.getAllFiles();
+    const allFilenames = new Set(allFiles.map((f) => f.sourceFileName));
+    if (allFilenames.has(fileName)) {
+      throw new BadRequestException({
+        message: 'Invalid filename, this already exists',
+      });
+    }
+
+    // Throws an error if no rules exist for the specified program
+    const rules = await this.alertService.getRulesForProgram(program);
+    if (!rules) {
+      throw new HttpException(
+        `No rules established for program ${program}`,
+        HttpStatus.FORBIDDEN
+      );
+    }
+
+    // Creates a new daily status for the rule, if none exist, so that files can be tracked
+    let daily = await this.alertService.getDailyForRule(rules, new Date());
+    if (!daily) {
+      daily = await this.alertService.createNewDaily(rules, new Date());
+    }
+
+    try {
+      // FileType is based on the filename (from Parser) or from the endpoint body
+      if (fileType === FileTypes.SBC_SALES) {
+        this.appLogger.log('Parse and store SBC Sales in DB...', fileName);
+        const garmsSales = await this.parseGarmsFile(contents, fileName); // validating step
+        const fileToSave = await this.saveFileUploaded({
+          sourceFileType: fileType,
+          sourceFileName: fileName,
+          sourceFileLength: garmsSales.length,
+          dailyUpload: daily,
+        });
+        this.appLogger.log(`Transaction count: ${garmsSales.length}`);
+        await this.transactionService.saveTransactions(
+          garmsSales.map((sale) => ({
+            ...sale,
+            fileUploadedEntityId: fileToSave.id,
+          }))
+        );
+      }
+
+      if (fileType === FileTypes.TDI17) {
+        this.appLogger.log('Parse and store TDI17 in DB...', fileName);
+        const cashDeposits = await this.parseTDICashFile(
+          fileName,
+          program,
+          buffer
+        ); // validating step
+        const fileToSave = await this.saveFileUploaded({
+          sourceFileType: fileType,
+          sourceFileName: fileName,
+          sourceFileLength: cashDeposits.length,
+          dailyUpload: daily,
+        });
+        this.appLogger.log(`Cash Deposits count: ${cashDeposits.length}`);
+        await this.cashDepositService.saveCashDepositEntities(
+          cashDeposits.map((deposit) => ({
+            ...deposit,
+            fileUploadedEntityId: fileToSave.id,
+          }))
+        );
+      }
+
+      if (fileType === FileTypes.TDI34) {
+        this.appLogger.log('Parse and store TDI34 in DB...', fileName);
+        const posEntities = await this.parseTDICardsFile(
+          fileName,
+          program,
+          buffer
+        );
+        const fileToSave = await this.saveFileUploaded({
+          sourceFileType: fileType,
+          sourceFileName: fileName,
+          sourceFileLength: posEntities.length,
+          dailyUpload: daily,
+        });
+        this.appLogger.log(`POS Deposits count: ${posEntities.length}`);
+        await this.posDepositService.savePOSDepositEntities(
+          posEntities.map((deposit) => ({
+            ...deposit,
+            fileUploadedEntityId: fileToSave.id,
+            timestamp: deposit.timestamp,
+          }))
+        );
+      }
+    } catch (err: unknown) {
+      const message =
+        err instanceof Error
+          ? err.message
+          : `Error with processing ${fileName}`;
+      this.appLogger.log(message);
+      throw new BadRequestException({ message });
+    }
   }
 }
