+++ apps/backend/test/mocks/classes/file_ingestion_rules_mock.ts
@@ -1,7 +1,8 @@
 import { faker } from '@faker-js/faker';
-import { FileIngestionRulesEntity } from '../../../src/parse/entities/file-ingestion-rules.entity';
 import { ProgramRequiredFileMock } from './required_file_mock';
 import { FileTypes } from '../../../src/constants';
+import { FileIngestionRulesEntity } from '../../../src/notification/entities/file-ingestion-rules.entity';
+import { ProgramRequiredFileEntity } from '../../../src/parse/entities/program-required-file.entity';
 
 export class FileIngestionRulesMock extends FileIngestionRulesEntity {
   constructor(program: string) {
@@ -13,7 +14,11 @@ export class FileIngestionRulesMock extends FileIngestionRulesEntity {
   }
 
   public assignRequiredFile(filename: string, fileType: FileTypes) {
-    const requiredFile = new ProgramRequiredFileMock(filename, fileType, this);
+    const requiredFile = new ProgramRequiredFileMock(
+      filename,
+      fileType,
+      this
+    ) as ProgramRequiredFileEntity;
     this.requiredFiles.push(requiredFile);
   }
 }
