+++ terraform/lambda_dailyfilecheck.tf
@@ -1,11 +1,11 @@
-resource "aws_lambda_function" "parser" {
-  description                    = "Parser function ${local.namespace}"
-  function_name                  = "parser"
+resource "aws_lambda_function" "daily-alert" {
+  description                    = "Daily alert function ${local.namespace}"
+  function_name                  = "dailyFileCheck-${local.namespace}"
   role                           = aws_iam_role.lambda.arn
   runtime                        = "nodejs18.x"
   filename                       = "build/empty_lambda.zip"
   source_code_hash               = filebase64sha256("build/empty_lambda.zip")
-  handler                        = "src/lambdas/parser.handler"
+  handler                        = "src/lambdas/dailyfilecheck.handler"
   memory_size                    = 1024
   timeout                        = 600
   reserved_concurrent_executions = 1
@@ -30,33 +30,19 @@ resource "aws_lambda_function" "parser" {
     }
   }
 
-  lifecycle {
-    ignore_changes = [
-      # Ignore changes to tags, e.g. because a management agent
-      # updates these based on some ruleset managed elsewhere.
-      filename,
-      source_code_hash,
-    ]
-  }
+  # lifecycle {
+  #   ignore_changes = [
+  #     # Ignore changes to tags, e.g. because a management agent
+  #     # updates these based on some ruleset managed elsewhere.
+  #     filename,
+  #     source_code_hash,
+  #   ]
+  # }
 }
 
-resource "aws_s3_bucket_notification" "aws-lambda-trigger" {
-  bucket = "${aws_s3_bucket.sftp_storage.id}"
-  lambda_function {
-    lambda_function_arn = "${aws_lambda_function.parser.arn}"
-    events              = ["s3:ObjectCreated:*"]
-  }
-}
-
-resource "aws_lambda_permission" "trigger-parse" {
-  statement_id  = "AllowS3Invoke"
-  action        = "lambda:InvokeFunction"
-  function_name = "${aws_lambda_function.parser.function_name}"
-  principal = "s3.amazonaws.com"
-  source_arn = "arn:aws:s3:::${aws_s3_bucket.sftp_storage.id}"
-}
-
-output "arn" {
-  value = "${aws_lambda_function.parser.arn}"
-}
+# resource "aws_cloudwatch_event_rule" "daily-alert" {
+#   name                = "daily-alert"
+#   description         = "Fires every day at 11 AM and 3 PM"
+#   schedule_expression = "cron(0 11,15/12 * * ? * )"
+# }
